requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
langchain==0.0.350
langchain-google-genai==0.0.5
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
scikit-learn>=1.0.0
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas>=2.0.0
numpy>=1.24.0
streamlit-ace==0.1.1
streamlit-option-menu==0.3.6
firebase-admin==6.2.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
pyperclip==1.8.2 